-Update the USDA nutrient database to SR 16-1 -Let the user create a new category from the dialog to select a category -Give the user a warning before deleting all recipes associated with a prep method when that prep method is removed. -Reload the ingredient matcher ingredients after importing. -Set the 'Cancel' button in the DependenciesDialog to be the default button for safety to our users ;)
svn path=/trunk/kdeextragear-3/krecipes/; revision=296029

+++ src/dialogs/prepmethodsdialog.cpp
@@ -12,6 +12,7 @@
 
 #include "prepmethodsdialog.h"
 #include "createelementdialog.h"
+#include "dependanciesdialog.h"
 #include "DBBackend/recipedb.h"
 
 #include <klocale.h>
@@ -104,11 +105,18 @@ if ( (it=prepMethodListView->selectedItem()) ) prepMethodID=it->text(0).toInt();
 
 if (prepMethodID>=0) // an prepMethod was selected previously
 {
-database->removePrepMethod(prepMethodID);
+	ElementList dependingRecipes;
+	database->findPrepMethodDependancies(prepMethodID,&dependingRecipes);
+	if (dependingRecipes.isEmpty()) database->removePrepMethod(prepMethodID);
+	else // Need Warning! 
+	{
+		DependanciesDialog *warnDialog = new DependanciesDialog(this,&dependingRecipes);
+		if (warnDialog->exec()==QDialog::Accepted) database->removePrepMethod(prepMethodID);
+		delete warnDialog;
+	}
 }
 
 reload();// Reload the list from the database
-
 }
 
 void PrepMethodsDialog::modPrepMethod(QListViewItem* i)
